backend/api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
boto3>=1.26.0
neo4j>=5.0.0

# Strands Agents dependencies
strands-agents>=0.2.0
anthropic>=0.18.0
